requirements.txt
alembic==1.8.1
anyio==3.6.1
autopep8==1.7.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
greenlet==1.1.3.post0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
orjson==3.8.0
passlib==1.7.4
psycopg2==2.9.4
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.23
starlette==0.20.4
toml==0.10.2
typing_extensions==4.4.0
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
